+++ sshd-sftp/src/test/java/org/apache/sshd/sftp/client/fs/SftpFileSystemTest.java
@@ -57,6 +57,7 @@ import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.TreeMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -97,16 +98,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNotSame;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
@@ -701,6 +692,41 @@ public class SftpFileSystemTest extends AbstractSftpFilesSystemSupport {
         assertArrayEquals(expected, actual, "Mismatched persisted data");
         actual = Files.readAllBytes(lclFile2);
         assertArrayEquals(expected, actual, "Mismatched copied data");
+    }
+
+    @Test
+    void fileCopyInChunks() throws IOException {
+        Path targetPath = detectTargetFolder();
+        Path lclSftp = CommonTestSupportUtils.resolve(targetPath, SftpConstants.SFTP_SUBSYSTEM_NAME,
+                getClass().getSimpleName());
+        Files.createDirectories(lclSftp);
+
+        Path lclFile = lclSftp.resolve(getCurrentTestName() + ".txt");
+        Files.deleteIfExists(lclFile);
+        Path lclFile2 = lclSftp.resolve(getCurrentTestName() + ".txt2");
+        Files.deleteIfExists(lclFile2);
+        byte[] expected = new byte[256 * 1024 + 200];
+        Random rnd = new Random();
+        rnd.nextBytes(expected);
+        Files.write(lclFile, expected);
+        try (FileSystem fs = FileSystems.newFileSystem(createDefaultFileSystemURI(), Collections.emptyMap())) {
+            Path parentPath = targetPath.getParent();
+            Path remotePath = fs.getPath(CommonTestSupportUtils.resolveRelativeRemotePath(parentPath, lclFile));
+
+            try (FileChannel sourceChannel = FileChannel.open(remotePath, StandardOpenOption.READ);
+                 FileChannel targetChannel = FileChannel.open(lclFile2, StandardOpenOption.CREATE, StandardOpenOption.WRITE,
+                         StandardOpenOption.TRUNCATE_EXISTING)) {
+
+                long totalBytes = sourceChannel.size();
+                long copiedBytes = 0L;
+
+                while (copiedBytes < totalBytes) {
+                    copiedBytes += sourceChannel.transferTo(targetChannel.position(), /* count = */ 1024, targetChannel);
+                }
+            }
+            byte[] actual = Files.readAllBytes(lclFile2);
+            assertArrayEquals(expected, actual, "Mismatched read data");
+        }
     }
 
     @Test